Refactor API method test framework
