+++ api/endpoints/contact/index.spec.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var ContactAPI = require('./');
-var testApiMethod = require('../../../test-api-method');
+var testApiMethod = require('../_test');
 
 describe('SuiteAPI Contact endpoint', function() {
