+++ api/endpoints/segment/index.spec.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var SegmentAPI = require('./');
-var testApiMethod = require('../../../test-api-method');
+var testApiMethod = require('../_test');
 
 describe('SuiteAPI Segment endpoint', function() {
