+++ api/endpoints/settings/index.spec.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var SettingsAPI = require('./');
-var testApiMethod = require('../../../test-api-method');
+var testApiMethod = require('../_test');
 
 describe('SuiteAPI Settings endpoint', function() {
